Fixed issue so more than one recipe can be created. Added feature to edit an existing recipe.

+++ db/schema.sql
@@ -14,12 +14,12 @@ CREATE TABLE users (
 );
 
 CREATE TABLE savedRecipes (
-	id SERIAL PRIMARY KEY,
+	id BIGSERIAL PRIMARY KEY,
 	name VARCHAR,
 	measurements VARCHAR,
 	ingredients VARCHAR,
 	instructions TEXT,
 	image TEXT,
 	beverageType VARCHAR,
-	user_id SERIAL REFERENCES users(id)
+	user_id BIGINT REFERENCES users(id)
 );
